packages/api-backend/requirements.txt
fastapi[all]
uvicorn
python-dotenv
redis
langgraph
langchain
langchain-cohere
passlib[bcrypt]